requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.10.1
google-auth==2.35.0
python-jose[cryptography]==3.3.0
